% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePerm.R
\name{combResults}
\alias{combResults}
\title{Combine results obtained from separate workers}
\usage{
combResults(results, x)
}
\arguments{
\item{results}{list of combined results}

\item{x}{list of results from one worker}
}
\value{
new list of combined results
}
\description{
Combine results obtained from separate workers
}
