% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractShapleyfromLog.R
\name{extractShapley}
\alias{extractShapley}
\title{Extract Shapley data from temporary savings}
\usage{
extractShapley(fname)
}
\arguments{
\item{fname}{temporary save file to extract data from}
}
\value{
data.frame having the same column names as properties of shapley class.
}
\description{
Extract Shapley data from temporary savings
}
