% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataShapley.R
\name{dataShapleyI5.MT}
\alias{dataShapleyI5.MT}
\title{Data Shapley with different truncation rule (multithreading)}
\usage{
dataShapleyI5.MT(
  D,
  A,
  V,
  T,
  tol = 0.01,
  convTol = tol * 5,
  log.file = "",
  log.append = F,
  rdata.name = "tmpShapley",
  doToleranceBreak = FALSE,
  cluster.size = 4
)
}
\arguments{
\item{D}{dataset of N points}

\item{A}{learning algorithm}

\item{V}{performance score}

\item{T}{test dataset}

\item{tol}{trancation tolerance}

\item{convTol}{convergence tolerance}

\item{log.file}{name of file to be used in `cat` calls}

\item{log.append}{logical parameter indicating if `cat` should append its output to log-file}

\item{rdata.name}{name of RData file should be used to save intermediate calculation results}
}
\value{
Shapley value of training points
}
\description{
Data Shapley with different truncation rule (multithreading)
}
