% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataShapley.R
\name{dataShapleyI5.MT}
\alias{dataShapleyI5.MT}
\title{Data Shapley with different truncation rule (multithreading)}
\usage{
dataShapleyI5.MT(
  D,
  A,
  V,
  T,
  tol = 0.01,
  convTol = tol * 5,
  log.file = "",
  log.append = F,
  rdata.name = "tmpShapleyML",
  cluster.size = 4,
  conv_check_step = 100,
  base.seed = as.numeric(Sys.time()),
  .continue = TRUE
)
}
\arguments{
\item{D}{dataset of N points}

\item{A}{learning algorithm}

\item{V}{performance score}

\item{T}{test dataset}

\item{tol}{truncation tolerance}

\item{convTol}{convergence tolerance}

\item{log.file}{name of file to be used in `cat` calls}

\item{log.append}{logical parameter indicating if `cat` should append its output to log-file}

\item{rdata.name}{name of RData file should be used to save intermediate calculation results}

\item{cluster.size}{number of CPU cores to be used for multiparallel computing}

\item{conv_check_step}{convergence is to be computed and checked every this number of iterations}

\item{base.seed}{Seed value to be used as basic value in different workers}
}
\value{
Shapley value of training points
}
\description{
Data Shapley with different truncation rule (multithreading)
}
